Remove `AttributeLintKind::MalFormedDiagnosticAttribute` variant
